+++ autosweep/stream.cc
@@ -6,8 +6,8 @@ std::ostream& operator<<(std::ostream& stream, const Cell& cell) {
 }
 
 std::ostream& operator<<(std::ostream& stream, const Board& board) {
-  for (int i = 0; i < board.rows(); ++i) {
-    for (int j = 0; j < board.cols(); ++j) {
+  for (size_t i = 0; i < board.rows(); ++i) {
+    for (size_t j = 0; j < board.cols(); ++j) {
       stream << board.at(i, j);
     }
     stream << '\n';
